Rename issueeditor to IssueRuleEditor
