+++ src/sentry/static/sentry/app/views/settings/projectAlerts/issueRuleEditor/index.tsx
