+++ src/sentry/static/sentry/app/views/settings/projectAlerts/issueRuleEditor/ticketRuleModal.tsx
